Rename `Traits` to `Steps`
